Introducing Json Format
The Json format writes a complete Json document for each test case.
The results for test cases are new-line separated.

This commit adds all the filed that are in the Html and Text tables,
but the intention is to be able to get full stack traces and still
have the ability to filter line-by-line using grep etc.